+++ src/main/java/org/averyocean65/folklore/vectors/Vector2.java
@@ -1,33 +1,33 @@
 package org.averyocean65.folklore.vectors;
 
 public class Vector2 implements VectorBase<Vector2> {
-    private float x;
-    private float y;
+    private double x;
+    private double y;
 
-    public Vector2(float x, float y) {
+    public Vector2(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
     @Override
-    public float getMagnitude() {
-        return (float)Math.sqrt(x*x + y*y);
+    public double getMagnitude() {
+        return Math.sqrt(x*x + y*y);
     }
 
     @Override
-    public Vector2 getScaled(float factor) {
+    public Vector2 getScaled(double factor) {
         return null;
     }
 
     @Override
     public Vector2 getNormalized() {
-        float newX = x / getMagnitude();
-        float newY = y / getMagnitude();
+        double newX = x / getMagnitude();
+        double newY = y / getMagnitude();
         return new Vector2(newX, newY);
     }
 
     @Override
-    public float[] toArray() {
-        return new float[] { x, y };
+    public double[] toArray() {
+        return new double[] { x, y };
     }
 }
