+++ src/main/java/org/averyocean65/folklore/vectors/VectorBase.java
@@ -1,8 +1,8 @@
 package org.averyocean65.folklore.vectors;
 
 public interface VectorBase<T> {
-    public float getMagnitude();
-    public T getScaled(float factor);
+    public double getMagnitude();
+    public T getScaled(double factor);
     public T getNormalized();
-    public float[] toArray();
+    public double[] toArray();
 }
